requirements.txt
openai
psycopg2
python-dotenv
scikit-learn
